refactor(useDefindex): extract vault metadata logic to hooks/use-vault-meta.ts
Step 2/4 of #66. Moves MercatoVaultMeta type, vaultMeta/vaultInfoError
state, fetchVaultMeta, and the mount effect into a dedicated hook.
useDefindex.ts now composes useVaultMeta() and re-exports the type so
existing consumers and lib/defindex/vault-cache.ts need no changes.

Verified: tsc --noEmit error count unchanged (10, same as baseline).
No consumer files modified. Cross-checked with cline.
